refactor(call-taker): move components to files, fix date/time
