+++ lib/components/form/call-taker/advanced-options.js
@@ -1,0 +1,228 @@
+import { hasBike } from '@opentripplanner/core-utils/lib/itinerary'
+import {SubmodeSelector} from '@opentripplanner/trip-form'
+import * as TripFormClasses from '@opentripplanner/trip-form/lib/styled'
+import React, { Component } from 'react'
+import Select from 'react-select'
+import styled from 'styled-components'
+
+import { modeButtonButtonCss } from '../styled'
+import { ComponentContext } from '../../../util/contexts'
+
+export const StyledSubmodeSelector = styled(SubmodeSelector)`
+  ${TripFormClasses.SubmodeSelector.Row} {
+    > * {
+      padding: 3px 5px 3px 0px;
+    }
+    > :last-child {
+      padding-right: 0px;
+    }
+    ${TripFormClasses.ModeButton.Button} {
+      padding: 6px 12px;
+    }
+    svg,
+    img {
+      margin-left: 0px;
+    }
+  }
+  ${TripFormClasses.SubmodeSelector.InlineRow} {
+    margin: -3px 0px;
+  }
+
+  ${TripFormClasses.SubmodeSelector} {
+      ${modeButtonButtonCss}
+  }
+`
+
+const metersToMiles = meters => Math.round(meters * 0.000621371 * 100) / 100
+const milesToMeters = miles => miles / 0.000621371
+
+export default class AdvancedOptions extends Component {
+  constructor (props) {
+    super(props)
+    this.state = {
+      expandAdvanced: props.expandAdvanced,
+      routeOptions: [],
+      transitModes: props.modes.transitModes.map(m => m.mode)
+    }
+  }
+
+  static contextType = ComponentContext
+
+  componentWillMount () {
+    // Fetch routes for banned/preferred routes selectors.
+    this.props.findRoutes()
+  }
+
+  componentWillReceiveProps (nextProps) {
+    const {routes} = nextProps
+    // Once routes are available, map them to the route options format.
+    if (routes && !this.props.routes) {
+      const routeOptions = Object.values(routes).map(this.routeToOption)
+      this.setState({routeOptions})
+    }
+  }
+
+  _setBannedRoutes = options => {
+    const bannedRoutes = options ? options.map(o => o.value).join(',') : ''
+    this.props.setQueryParam({ bannedRoutes })
+  }
+
+  _setPreferredRoutes = options => {
+    const preferredRoutes = options ? options.map(o => (o.value)).join(',') : ''
+    this.props.setQueryParam({ preferredRoutes })
+  }
+
+  _isBannedRouteOptionDisabled = option => {
+    // Disable routes that are preferred already.
+    const preferredRoutes = this.getRouteList('preferredRoutes')
+    return preferredRoutes && preferredRoutes.find(o => o.value === option.value)
+  }
+
+  _isPreferredRouteOptionDisabled = option => {
+    // Disable routes that are banned already.
+    const bannedRoutes = this.getRouteList('bannedRoutes')
+    return bannedRoutes && bannedRoutes.find(o => o.value === option.value)
+  }
+
+  getDistanceStep = distanceInMeters => {
+    // Determine step for max walk/bike based on current value. Increment by a
+    // quarter mile if dealing with small values, whatever number will round off
+    // the number if it is not an integer, or default to one mile.
+    return metersToMiles(distanceInMeters) <= 2
+      ? '.25'
+      : metersToMiles(distanceInMeters) % 1 !== 0
+        ? `${metersToMiles(distanceInMeters) % 1}`
+        : '1'
+  }
+
+  _onSubModeChange = changedMode => {
+    // Get previous transit modes from state and all modes from query.
+    const transitModes = [...this.state.transitModes]
+    const allModes = this.props.currentQuery.mode.split(',')
+    const index = transitModes.indexOf(changedMode)
+    if (index === -1) {
+      // If submode was not selected, add it.
+      transitModes.push(changedMode)
+      allModes.push(changedMode)
+    } else {
+      // Otherwise, remove it.
+      transitModes.splice(index, 1)
+      const i = allModes.indexOf(changedMode)
+      allModes.splice(i, 1)
+    }
+    // Update transit modes in state.
+    this.setState({transitModes})
+    // Update all modes in query (set to walk if all transit modes inactive).
+    this.props.setQueryParam({ mode: allModes.join(',') || 'WALK' })
+  }
+
+  _setMaxWalkDistance = evt => {
+    this.props.setQueryParam({ maxWalkDistance: milesToMeters(evt.target.value) })
+  }
+
+  /**
+   * Get list of routes for specified key (either 'bannedRoutes' or
+   * 'preferredRoutes').
+   */
+  getRouteList = key => {
+    const routesParam = this.props.currentQuery[key]
+    const idList = routesParam ? routesParam.split(',') : []
+    if (this.state.routeOptions) {
+      return this.state.routeOptions.filter(o => idList.indexOf(o.value) !== -1)
+    } else {
+      // If route list is not available, default labels to route IDs.
+      return idList.map(id => ({value: id, label: id}))
+    }
+  }
+
+  routeToOption = route => {
+    if (!route) return null
+    const {id, longName, shortName} = route
+    // For some reason the OTP API expects route IDs in this double
+    // underscore format
+    // FIXME: This replace is flimsy! What if there are more colons?
+    const value = id.replace(':', '__')
+    const label = shortName
+      ? `${shortName}${longName ? ` - ${longName}` : ''}`
+      : longName
+    return {value, label}
+  }
+
+  render () {
+    const { currentQuery, modes, onKeyDown } = this.props
+    const { ModeIcon } = this.context
+
+    const {maxBikeDistance, maxWalkDistance, mode} = currentQuery
+    const bannedRoutes = this.getRouteList('bannedRoutes')
+    const preferredRoutes = this.getRouteList('preferredRoutes')
+    const transitModes = modes.transitModes.map(modeObj => {
+      const modeStr = modeObj.mode || modeObj
+      return {
+        id: modeStr,
+        selected: this.state.transitModes.indexOf(modeStr) !== -1,
+        text: (
+          <span>
+            <ModeIcon mode={modeStr} />
+          </span>
+        ),
+        title: modeObj.label
+      }
+    })
+    // FIXME: Set units via config.
+    const unitsString = '(mi.)'
+    return (
+      <div>
+        <div style={{alignItems: 'center', display: 'flex', justifyContent: 'space-between'}}>
+          <label style={{fontWeight: '400'}}>
+            Max walk {unitsString}
+            <input
+              onChange={this._setMaxWalkDistance}
+              onKeyDown={onKeyDown}
+              step={this.getDistanceStep(maxWalkDistance)}
+              min='0'
+              style={{display: 'block', marginRight: '10px', width: '60px'}}
+              value={metersToMiles(maxWalkDistance)}
+              type='number' />
+          </label>
+          {hasBike(mode)
+            ? <label style={{fontWeight: '400'}}>
+              Max bike {unitsString}
+              <input
+                onChange={this._setMaxBikeDistance}
+                onKeyDown={onKeyDown}
+                step={this.getDistanceStep(maxBikeDistance)}
+                min='0'
+                style={{display: 'block', marginRight: '10px', width: '60px'}}
+                value={metersToMiles(maxBikeDistance)}
+                type='number' />
+            </label>
+            : null
+          }
+          <StyledSubmodeSelector
+            modes={transitModes}
+            onChange={this._onSubModeChange}
+            // FIXME: Need to pass onKeyDown to children buttons in
+            // otp-ui.
+            onKeyDown={onKeyDown}
+          />
+        </div>
+        <Select
+          value={preferredRoutes}
+          id='preferredRoutes'
+          isMulti
+          isOptionDisabled={this._isPreferredRouteOptionDisabled}
+          options={this.state.routeOptions}
+          onChange={this._setPreferredRoutes}
+          placeholder='Select preferred routes...' />
+        <Select
+          value={bannedRoutes}
+          id='bannedRoutes'
+          isMulti
+          isOptionDisabled={this._isBannedRouteOptionDisabled}
+          options={this.state.routeOptions}
+          onChange={this._setBannedRoutes}
+          placeholder='Select banned routes...' />
+      </div>
+    )
+  }
+}
